+++ package.json
@@ -16,6 +16,7 @@
     "vue": "^2.6.11",
     "vue-class-component": "^7.2.2",
     "vue-property-decorator": "^8.3.0",
+    "vue-router": "^3.1.6",
     "vue-suggestion": "^1.1.0",
     "vue-tippy": "^4.4.0",
     "vuebar": "0.0.20",
